feat: allow battery replacement caused resets
Remove the check for the highest counter value, as devices may restart the sequence for various reasons. A new security counter value will be accepted only if decryption is successful, indicating that the sender is either a legitimate user with the key, or is replaying BLE messages since the last setup or device reset. This requires an attacker to be extremely patient and dedicated, with questionable rewards.